requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.3
idna==3.2
numpy==1.21.1
opencv-python==4.5.3.56
Pillow==8.3.1
psycopg2-binary==2.9.1
pyTelegramBotAPI==3.8.2
qrcode==7.2
requests==2.26.0
urllib3==1.26.6
